# recipesshop/app/Http/Controllers/RecipeController.php
<?php

namespace App\Http\Controllers;

use App\Http\Resources\IngredientResource;
use App\Http\Resources\RecipeResource;
use App\Models\Ingredient;
use App\Models\Recipe;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Auth;
use Illuminate\Support\Facades\Validator;
use Illuminate\Support\Str;

class RecipeController extends Controller
{
     /**
     * @OA\Get(
     *   path="/api/recipes",
     *   tags={"Recipes"},
     *   summary="List recipes (with search, filters, sort, pagination)",
     *   @OA\Parameter(name="search", in="query", required=false, @OA\Schema(type="string"), description="Search in name, description and ingredient names"),
     *   @OA\Parameter(name="ingredients_any", in="query", required=false, @OA\Schema(type="string"), description="CSV of ingredient IDs; recipe must contain ANY of them"),
     *   @OA\Parameter(name="ingredients_all", in="query", required=false, @OA\Schema(type="string"), description="CSV of ingredient IDs; recipe must contain ALL of them"),
     *   @OA\Parameter(name="ingredients_exclude", in="query", required=false, @OA\Schema(type="string"), description="CSV of ingredient IDs to exclude"),
     *   @OA\Parameter(
     *     name="sort", in="query", required=false,
     *     @OA\Schema(type="string", enum={"name","-name","created_at","-created_at","updated_at","-updated_at","ingredients_count","-ingredients_count"}),
     *     description="Sort field (prefix with - for DESC)"
     *   ),
     *   @OA\Parameter(name="per_page", in="query", required=false, @OA\Schema(type="integer", minimum=1, maximum=100), description="Items per page (default 15)"),
     *   @OA\Parameter(name="page", in="query", required=false, @OA\Schema(type="integer", minimum=1), description="Page number (default 1)"),
     *   @OA\Response(
     *     response=200,
     *     description="OK",
     *     @OA\JsonContent(
     *       type="object",
     *       @OA\Property(property="meta", type="object",
     *         @OA\Property(property="page", type="integer", example=1),
     *         @OA\Property(property="per_page", type="integer", example=15),
     *         @OA\Property(property="total", type="integer", example=42),
     *         @OA\Property(property="last_page", type="integer", example=3)
     *       ),
     *       @OA\Property(property="recipes", type="array",
     *         @OA\Items(type="object",
     *           @OA\Property(property="id", type="integer", example=7),
     *           @OA\Property(property="name", type="string", example="Greek Salad"),
     *           @OA\Property(property="description", type="string", example="Fresh and easy."),
     *           @OA\Property(property="ingredient_ids", type="array", @OA\Items(type="integer"), example={1,2,3,5}),
     *           @OA\Property(property="ingredients", type="array",
     *             @OA\Items(type="object",
     *               @OA\Property(property="id", type="integer", example=1),
     *               @OA\Property(property="name", type="string", example="Tomato"),
     *               @OA\Property(property="price", type="number", format="float", example=1.20)
     *             )
     *           )
     *         )
     *       )
     *     )
     *   ),
     *   @OA\Response(response=404, description="No recipes found.")
     * )
     */
    public function index(Request $request)
    {
        $v = Validator::make($request->all(), [
            'search' => ['sometimes', 'string', 'max:200'],
            'ingredients_any' => ['sometimes', 'string'],
            'ingredients_all' => ['sometimes', 'string'],
            'ingredients_exclude' => ['sometimes', 'string'],
            'sort' => ['sometimes', 'string', 'in:name,-name,created_at,-created_at,updated_at,-updated_at,ingredients_count,-ingredients_count'],
            'per_page' => ['sometimes', 'integer', 'min:1', 'max:100'],
            'page' => ['sometimes', 'integer', 'min:1'],
        ]);
        $v->validate();

        $perPage = (int) $request->input('per_page', 15);
        $page    = (int) $request->input('page', 1);
        $search  = trim((string) $request->input('search', ''));
        $sort    = (string) $request->input('sort', 'name');

        $parseIds = fn($csv) => array_values(array_unique(
            array_filter(array_map('intval', explode(',', (string) $csv)), fn($i) => $i > 0)
        ));

        $idsAny = $parseIds($request->input('ingredients_any'));
        $idsAll = $parseIds($request->input('ingredients_all'));
        $idsExclude = $parseIds($request->input('ingredients_exclude'));

        $q = Recipe::query();

        if ($search !== '') {
            $escaped = str_replace(['%', '_'], ['\%', '\_'], $search);

            $matchedIngredientIds = Ingredient::query()
                ->where('name', 'like', "%{$escaped}%")
                ->pluck('id')
                ->all();

            $q->where(function ($w) use ($search, $matchedIngredientIds) {
                $w->where('name', 'like', "%{$search}%")
                    ->orWhere('description', 'like', "%{$search}%");

                if (!empty($matchedIngredientIds)) {
                    $w->orWhere(function ($ww) use ($matchedIngredientIds) {
                        foreach ($matchedIngredientIds as $id) {
                            $ww->orWhereRaw('JSON_CONTAINS(ingredient_ids, ?)', [json_encode($id)]);
                        }
                    });
                }
            });
        }

        if (!empty($idsAny)) {
            $q->where(function ($w) use ($idsAny) {
                foreach ($idsAny as $id) {
                    $w->orWhereRaw('JSON_CONTAINS(ingredient_ids, ?)', [json_encode($id)]);
                }
            });
        }

        if (!empty($idsAll)) {
            foreach ($idsAll as $id) {
                $q->whereRaw('JSON_CONTAINS(ingredient_ids, ?)', [json_encode($id)]);
            }
        }

        if (!empty($idsExclude)) {
            foreach ($idsExclude as $id) {
                $q->whereRaw('NOT JSON_CONTAINS(ingredient_ids, ?)', [json_encode($id)]);
            }
        }

        $direction = Str::startsWith($sort, '-') ? 'desc' : 'asc';
        $field = ltrim($sort, '-');

        switch ($field) {
            case 'name':
            case 'created_at':
            case 'updated_at':
                $q->orderBy($field, $direction);
                break;
            case 'ingredients_count':
                $q->orderByRaw('JSON_LENGTH(ingredient_ids) ' . $direction);
                break;
            default:
                $q->orderBy('name', 'asc');
        }

        $recipes = $q->paginate($perPage, ['*'], 'page', $page);

        if ($recipes->isEmpty()) {
            return response()->json('No recipes found.', 404);
        }

        return response()->json([
            'meta'    => [
                'page' => $recipes->currentPage(),
                'per_page' => $recipes->perPage(),
                'total' => $recipes->total(),
                'last_page' => $recipes->lastPage(),
            ],
            'recipes' => RecipeResource::collection($recipes),
        ]);
    }
    /**
     * @OA\Get(
     *   path="/api/recipes/{recipe}/ingredients",
     *   tags={"Recipes"},
     *   summary="Get the ingredients for a specific recipe",
     *   @OA\Parameter(
     *     name="recipe", in="path", required=true, description="Recipe ID",
     *     @OA\Schema(type="integer")
     *   ),
     *   @OA\Response(
     *     response=200,
     *     description="OK",
     *     @OA\JsonContent(
     *       type="object",
     *       @OA\Property(property="recipe_id", type="integer", example=5),
     *       @OA\Property(property="ingredients", type="array",
     *         @OA\Items(type="object",
     *           @OA\Property(property="id", type="integer", example=1),
     *           @OA\Property(property="name", type="string", example="Tomato"),
     *           @OA\Property(property="price", type="number", format="float", example=1.20)
     *         )
     *       )
     *     )
     *   ),
     *   @OA\Response(response=404, description="No ingredients found for this recipe.")
     * )
     */
    public function ingredients(Recipe $recipe)
    {
        $ids = $recipe->ingredient_ids ?? [];

        if (empty($ids)) {
            return response()->json('No ingredients found for this recipe.', 404);
        }

        $ingredients = Ingredient::whereIn('id', $ids)->get();
        $ingredients = $ingredients->sortBy(function ($ing) use ($ids) {
            return array_search($ing->id, $ids, true);
        })->values();

        return response()->json([
            'recipe_id'   => $recipe->id,
            'ingredients' => IngredientResource::collection($ingredients),
        ]);
    }

    /**
     * Show the form for creating a new resource.
     */
    public function create()
    {
        //
    }

    /**
     * @OA\Post(
     *   path="/api/recipes",
     *   tags={"Recipes"},
     *   summary="Create a new recipe (admin only)",
     *   security={{"bearerAuth":{}}},
     *   @OA\RequestBody(
     *     required=true,
     *     @OA\JsonContent(
     *       required={"name","ingredient_ids"},
     *       @OA\Property(property="name", type="string", maxLength=255, example="Greek Salad"),
     *       @OA\Property(property="description", type="string", example="Fresh and easy."),
     *       @OA\Property(property="ingredient_ids", type="array", @OA\Items(type="integer"), example={1,2,3,5})
     *     )
     *   ),
     *   @OA\Response(
     *     response=201,
     *     description="Recipe created",
     *     @OA\JsonContent(
     *       type="object",
     *       @OA\Property(property="message", type="string", example="Recipe created successfully"),
     *       @OA\Property(property="recipe",
     *         type="object",
     *         @OA\Property(property="id", type="integer", example=12),
     *         @OA\Property(property="name", type="string", example="Greek Salad"),
     *         @OA\Property(property="description", type="string", example="Fresh and easy."),
     *         @OA\Property(property="ingredient_ids", type="array", @OA\Items(type="integer"), example={1,2,3,5})
     *       )
     *     )
     *   ),
     *   @OA\Response(response=403, description="Only admins can create recipes"),
     *   @OA\Response(response=422, description="Validation error")
     * )
     */
    public function store(Request $request)
    {
        if (Auth::user()->role !== 'admin') {
            return response()->json(['error' => 'Only admins can create recipes'], 403);
        }

        $validated = $request->validate([
            'name' => 'required|string|max:255|unique:recipes,name',
            'description' => 'nullable|string',
            'ingredient_ids' => 'required|array|min:1',
            'ingredient_ids.*' => 'integer|distinct|exists:ingredients,id',
        ]);

        $validated['ingredient_ids'] = array_values(array_unique($validated['ingredient_ids']));

        $recipe = Recipe::create($validated);

        return response()->json([
            'message' => 'Recipe created successfully',
            'recipe' => new RecipeResource($recipe),
        ], 201);
    }
    /**
     * @OA\Get(
     *   path="/api/recipes/{recipe}",
     *   tags={"Recipes"},
     *   summary="Get a single recipe",
     *   @OA\Parameter(
     *     name="recipe", in="path", required=true, description="Recipe ID",
     *     @OA\Schema(type="integer")
     *   ),
     *   @OA\Response(
     *     response=200,
     *     description="OK",
     *     @OA\JsonContent(
     *       type="object",
     *       @OA\Property(property="recipe",
     *         type="object",
     *         @OA\Property(property="id", type="integer", example=7),
     *         @OA\Property(property="name", type="string", example="Greek Salad"),
     *         @OA\Property(property="description", type="string", example="Fresh and easy."),
     *         @OA\Property(property="ingredient_ids", type="array", @OA\Items(type="integer"), example={1,2,3,5})
     *       )
     *     )
     *   ),
     *   @OA\Response(response=404, description="Recipe not found")
     * )
     */
    public function show(Recipe $recipe)
    {
        return response()->json([
            'recipe' => new RecipeResource($recipe),
        ]);
    }

    /**
     * Show the form for editing the specified resource.
     */
    public function edit(Recipe $recipe)
    {
        //
    }

    /**
     * @OA\Put(
     *   path="/api/recipes/{recipe}",
     *   tags={"Recipes"},
     *   summary="Update a recipe (admin only)",
     *   security={{"bearerAuth":{}}},
     *   @OA\Parameter(
     *     name="recipe", in="path", required=true, description="Recipe ID",
     *     @OA\Schema(type="integer")
     *   ),
     *   @OA\RequestBody(
     *     required=false,
     *     @OA\JsonContent(
     *       @OA\Property(property="name", type="string", maxLength=255, example="Summer Greek Salad"),
     *       @OA\Property(property="description", type="string", example="With extra basil."),
     *       @OA\Property(property="ingredient_ids", type="array", @OA\Items(type="integer"), example={1,2,3,5,16})
     *     )
     *   ),
     *   @OA\Response(
     *     response=200,
     *     description="Recipe updated",
     *     @OA\JsonContent(
     *       type="object",
     *       @OA\Property(property="message", type="string", example="Recipe updated successfully"),
     *       @OA\Property(property="recipe",
     *         type="object",
     *         @OA\Property(property="id", type="integer", example=7),
     *         @OA\Property(property="name", type="string", example="Summer Greek Salad"),
     *         @OA\Property(property="description", type="string", example="With extra basil."),
     *         @OA\Property(property="ingredient_ids", type="array", @OA\Items(type="integer"), example={1,2,3,5,16})
     *       )
     *     )
     *   ),
     *   @OA\Response(response=403, description="Only admins can update recipes"),
     *   @OA\Response(response=422, description="Validation error")
     * )
     */
    public function update(Request $request, Recipe $recipe)
    {
        if (Auth::user()->role !== 'admin') {
            return response()->json(['error' => 'Only admins can update recipes'], 403);
        }

        $validated = $request->validate([
            'name' => 'sometimes|string|max:255|unique:recipes,name,' . $recipe->id,
            'description' => 'sometimes|nullable|string',
            'ingredient_ids' => 'sometimes|array|min:1',
            'ingredient_ids.*' => 'integer|distinct|exists:ingredients,id',
        ]);

        if (isset($validated['ingredient_ids'])) {
            $validated['ingredient_ids'] = array_values(array_unique($validated['ingredient_ids']));
        }

        $recipe->update($validated);

        return response()->json([
            'message' => 'Recipe updated successfully',
            'recipe' => new RecipeResource($recipe),
        ]);
    }

    /**
     * @OA\Delete(
     *   path="/api/recipes/{recipe}",
     *   tags={"Recipes"},
     *   summary="Delete a recipe (admin only)",
     *   security={{"bearerAuth":{}}},
     *   @OA\Parameter(
     *     name="recipe", in="path", required=true, description="Recipe ID",
     *     @OA\Schema(type="integer")
     *   ),
     *   @OA\Response(
     *     response=200,
     *     description="Recipe deleted",
     *     @OA\JsonContent(type="object", example={"message":"Recipe deleted successfully"})
     *   ),
     *   @OA\Response(response=403, description="Only admins can delete recipes")
     * )
     */
    public function destroy(Recipe $recipe)
    {
        if (Auth::user()->role !== 'admin') {
            return response()->json(['error' => 'Only admins can delete recipes'], 403);
        }

        $recipe->delete();

        return response()->json(['message' => 'Recipe deleted successfully']);
    }
}