# recipesshop/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API documentation for Recipes Web Shop.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "List recipes (with search, filters, sort, pagination)",
                "operationId": "8b9e556d4d747bd60232961951494bfb",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in name, description and ingredient names",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ingredients_any",
                        "in": "query",
                        "description": "CSV of ingredient IDs; recipe must contain ANY of them",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ingredients_all",
                        "in": "query",
                        "description": "CSV of ingredient IDs; recipe must contain ALL of them",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ingredients_exclude",
                        "in": "query",
                        "description": "CSV of ingredient IDs to exclude",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort field (prefix with - for DESC)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "-name",
                                "created_at",
                                "-created_at",
                                "updated_at",
                                "-updated_at",
                                "ingredients_count",
                                "-ingredients_count"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (default 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (default 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "recipes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 7
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Greek Salad"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Fresh and easy."
                                                    },
                                                    "ingredient_ids": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "integer"
                                                        },
                                                        "example": [
                                                            1,
                                                            2,
                                                            3,
                                                            5
                                                        ]
                                                    },
                                                    "ingredients": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Tomato"
                                                                },
                                                                "price": {
                                                                    "type": "number",
                                                                    "format": "float",
                                                                    "example": 1.2
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No recipes found."
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Create a new recipe (admin only)",
                "operationId": "46448289940e329a644b891be9c66943",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "ingredient_ids"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Greek Salad"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Fresh and easy."
                                    },
                                    "ingredient_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3,
                                            5
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe created successfully"
                                        },
                                        "recipe": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Greek Salad"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Fresh and easy."
                                                },
                                                "ingredient_ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer"
                                                    },
                                                    "example": [
                                                        1,
                                                        2,
                                                        3,
                                                        5
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only admins can create recipes"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/recipes/{recipe}/ingredients": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get the ingredients for a specific recipe",
                "operationId": "52a68b3239e5581273eff80aeef5345d",
                "parameters": [
                    {
                        "name": "recipe",
                        "in": "path",
                        "description": "Recipe ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "recipe_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "ingredients": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Tomato"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1.2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No ingredients found for this recipe."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}